requirements.txt
python-dotenv>=1.0.1
streamlit>=1.38.0
openai>=1.40.0

# LangChain ecossistema atualizado
langchain>=0.3.0
langchain-core>=0.3.3
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain_classic

# Vetorização e utilidades
faiss-cpu>=1.8.0
tiktoken>=0.7.0

# Processamento de PDF
PyMuPDF>=1.24.9
